+++ packages/cordova-plugin/src/web.ts
@@ -16,7 +16,11 @@ function getCurrentPosition(options: CurrentPositionOptions, success: (output: P
         accuracy: position.accuracy,
         heading: position.heading,
         speed: position.speed,
-        altitudeAccuracy: position.altitudeAccuracy
+        altitudeAccuracy: position.altitudeAccuracy,
+        magneticHeading: position.magneticHeading,
+        trueHeading: position.trueHeading,
+        headingAccuracy: position.headingAccuracy,
+        course: position.course
       },
       timestamp: position.timestamp,
     }